Extract user list into a fragment
