+++ app/build.gradle
@@ -38,7 +38,9 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.3.1'
     implementation 'com.google.android.material:material:1.4.0'
     implementation 'androidx.constraintlayout:constraintlayout:2.1.0'
-    implementation 'androidx.activity:activity-ktx:1.4.0-alpha01'
+    implementation 'androidx.activity:activity-ktx:1.4.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'androidx.fragment:fragment-ktx:1.4.0'
 
     // HTTP client.
     def retrofitVersion = "2.9.0"
@@ -56,7 +58,7 @@ dependencies {
     kapt "com.squareup.moshi:moshi-kotlin-codegen:$moshiVersion"
 
     // AndroidX.
-    def lifecycleVersion = "2.4.0-alpha03"
+    def lifecycleVersion = "2.4.0"
     implementation "androidx.lifecycle:lifecycle-runtime-ktx:$lifecycleVersion"
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:$lifecycleVersion"
