+++ app/src/main/java/com/vadimsalavatov/mobiledev/BaseViewModel.kt
@@ -1,0 +1,5 @@
+package com.vadimsalavatov.mobiledev
+
+import androidx.lifecycle.ViewModel
+
+abstract class BaseViewModel : ViewModel()
